APIHandlerを移植した
