+++ CharacterAlarm/Repository/APIRequest.swift
@@ -5,6 +5,19 @@ enum HttpMethod: String {
     case post = "POST"
 }
 
+struct APIHeader {
+    static let defaultHeader: [String: String] = ["Content-Type": "application/json"]
+
+    static func createAuthorizationRequestHeader(userID: String, authToken: String) -> [String: String] {
+        guard let authorization = "\(userID):\(authToken)".data(using: .utf8)?.base64EncodedString() else {
+            fatalError()
+        }
+        var requestHeader = Self.defaultHeader
+        requestHeader["Authorization"] = "Basic \(authorization)"
+        return requestHeader
+    }
+}
+
 struct APIRequest {
     static func createUrlRequest(baseUrl: String = environmentVariable.apiEndpoint, path: String, httpMethod: HttpMethod = .get, requestHeader: [String: String] = ["X-API-VERSION": "0", "Content-Type": "application/json"]) -> URLRequest {
         let url = URL(string: baseUrl + path)!
